+++ core/comps/conversations/srvc-message-model.js
@@ -55,7 +55,7 @@ angular.module('cmConversations')
             this.files = [];
             this.fileIds = [];
 
-            this.state = new cmStateManagement(['new','decrypted','loading', 'incomplete', 'sending', 'waitForFiles']);
+            this.state = new cmStateManagement(['new','decrypted','loading', 'incomplete', 'sending', 'waitForFiles','notified']);
 
             /**
              * Initialize Message Object
@@ -387,7 +387,6 @@ angular.module('cmConversations')
             };
 
             this.decryptFiles = function(passphrase){
-
                 angular.forEach(this.files, function(file, index){
                     if(file.state.is('onlyFileId')) {
                         file
@@ -432,7 +431,7 @@ angular.module('cmConversations')
             if(conversation != undefined && ('on' in conversation)) {
                 conversation.on('message:reInitFiles', function () {
                     if (self.state.is('incomplete')) {
-                        self.incompleteFiles.forEach(function (file) {
+                        self.incompleteFiles.forEach(function(file) {
                             file.importFile();
                         });
                     }